jordanElimination algorithm implemented

+++ src/main/java/FileOperator.java
@@ -11,15 +11,15 @@ public class FileOperator {
     private static int rows;
     private static int columns;
 
-    public static int[][] readBoard(Path p) throws IOException {
+    public static double[][] readBoard(Path p) throws IOException {
         if (Files.exists(p)) {
             List<String> lines = Files.readAllLines(p);
-            int[][] board = new int[lines.size()][lines.size() + 1];
+            double[][] board = new double[lines.size()][lines.size() + 1];
             int i = 0;
             int j = 0;
             for (String s : lines) {
                 for (String arg : s.split("\\s+")) {
-                    board[i][j] = Integer.parseInt(arg);
+                    board[i][j] = Double.parseDouble(arg);
                     j++;
                 }
                 j = 0;
